+++ Verilog.novaextension/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Version 0.0.2
 
 Added System Verilog to the supported extensions.
+Added basic linting support using verilator.
 
 ## Version 0.0.1